Move NavigationTimingData definition into metric module
